+++ packages/transducers/src/rfn/conj.ts
@@ -1,9 +1,6 @@
 import { Reducer } from "../api";
+import { reducer } from "../reduce";
 
 export function conj<T>(): Reducer<Set<T>, T> {
-    return [
-        () => new Set(),
-        (acc) => acc,
-        (acc, x) => acc.add(x),
-    ];
+    return reducer(() => new Set(), (acc, x) => acc.add(x));
 }