+++ packages/transducers/src/rfn/count.ts
@@ -1,9 +1,6 @@
 import { Reducer } from "../api";
+import { reducer } from "../reduce";
 
 export function count(offset = 0, step = 1): Reducer<number, any> {
-    return [
-        () => offset,
-        (acc) => acc,
-        (acc, _) => acc + step,
-    ];
+    return reducer(() => offset, (acc, _) => acc + step);
 }