+++ packages/transducers/src/rfn/push.ts
@@ -1,9 +1,6 @@
 import { Reducer } from "../api";
+import { reducer } from "../reduce";
 
 export function push<T>(): Reducer<T[], T> {
-    return [
-        () => [],
-        (acc) => acc,
-        (acc, x) => (acc.push(x), acc),
-    ];
+    return reducer(() => [], (acc, x) => (acc.push(x), acc));
 }