+++ packages/hdom/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.0.27"></a>
+## [3.0.27](https://github.com/thi-ng/umbrella/compare/@thi.ng/hdom@3.0.26...@thi.ng/hdom@3.0.27) (2018-07-04)
+
+
+
+
+**Note:** Version bump only for package @thi.ng/hdom
+
 <a name="3.0.26"></a>
 ## [3.0.26](https://github.com/thi-ng/umbrella/compare/@thi.ng/hdom@3.0.25...@thi.ng/hdom@3.0.26) (2018-06-21)