+++ packages/rstream/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.9.2"></a>
+## [1.9.2](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream@1.9.1...@thi.ng/rstream@1.9.2) (2018-07-09)
+
+
+
+
+**Note:** Version bump only for package @thi.ng/rstream
+
+<a name="1.9.1"></a>
+## [1.9.1](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream@1.9.0...@thi.ng/rstream@1.9.1) (2018-07-04)
+
+
+
+
+**Note:** Version bump only for package @thi.ng/rstream
+
+<a name="1.9.0"></a>
+# [1.9.0](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream@1.8.2...@thi.ng/rstream@1.9.0) (2018-07-04)
+
+
+### Features
+
+* **rstream:** add support for event listener opts in `fromEvent()` ([d5ac264](https://github.com/thi-ng/umbrella/commit/d5ac264))
+
+
+
+
+<a name="1.8.2"></a>
+## [1.8.2](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream@1.8.1...@thi.ng/rstream@1.8.2) (2018-07-03)
+
+
+
+
+**Note:** Version bump only for package @thi.ng/rstream
+
+<a name="1.8.1"></a>
+## [1.8.1](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream@1.8.0...@thi.ng/rstream@1.8.1) (2018-07-03)
+
+
+### Bug Fixes
+
+* **rstream:** unsubscribe on error in transduce() ([8c7d937](https://github.com/thi-ng/umbrella/commit/8c7d937))
+
+
+
+
 <a name="1.8.0"></a>
 # [1.8.0](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream@1.7.3...@thi.ng/rstream@1.8.0) (2018-06-21)