+++ packages/transducers-fsm/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.4"></a>
+## [0.2.4](https://github.com/thi-ng/umbrella/compare/@thi.ng/transducers-fsm@0.2.3...@thi.ng/transducers-fsm@0.2.4) (2018-07-09)
+
+
+
+
+**Note:** Version bump only for package @thi.ng/transducers-fsm
+
+<a name="0.2.3"></a>
+## [0.2.3](https://github.com/thi-ng/umbrella/compare/@thi.ng/transducers-fsm@0.2.2...@thi.ng/transducers-fsm@0.2.3) (2018-07-03)
+
+
+
+
+**Note:** Version bump only for package @thi.ng/transducers-fsm
+
 <a name="0.2.2"></a>
 ## [0.2.2](https://github.com/thi-ng/umbrella/compare/@thi.ng/transducers-fsm@0.2.1...@thi.ng/transducers-fsm@0.2.2) (2018-06-21)